+++ src/component/NavBar/NavBar.stories.jsx
