+++ src/constant/index.js
@@ -2,4 +2,6 @@ const FRUITS = ['🍅', '🍑', '🥝', '🥥'];
 
 const PRODUCTS_COUNT_PER_PAGE = 12;
 
-export { FRUITS, PRODUCTS_COUNT_PER_PAGE };
+const SNACKBAR_RUNNING_TIME = 2500;
+
+export { FRUITS, PRODUCTS_COUNT_PER_PAGE, SNACKBAR_RUNNING_TIME };
