+++ src/constant/messages.js
@@ -2,4 +2,8 @@ const ERROR_MESSAGE = {
   FAIL_TO_FETCH_PRODUCTS: '상품 목록을 불러오는 데에 실패하였습니다!',
 };
 
-export { ERROR_MESSAGE };
+const ALERT_MESSAGE = {
+  ADD_CART: '상품이 장바구니에 추가되었습니다.',
+};
+
+export { ERROR_MESSAGE, ALERT_MESSAGE };
